services/AccountService/requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.1.2
django-graphql-jwt==0.4.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
graphene==3.4
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
promise==2.3
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.0.4
